CRUD operations implementation

+++ app/src/main/java/training/journal/items/WorkoutItem.kt
@@ -4,7 +4,7 @@ import training.journal.items.interfaces.WithId
 
 data class WorkoutItem(
     override val id: String,
-    val iconId: Int,
+    val iconId: Long,
     val title: String,
     val description: String
 ) : WithId
